Service was added, command integration
