+++ .env
@@ -25,5 +25,5 @@ APP_SECRET=8191f3d8f24ca77e8bf523501a06893c
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
 # For a PostgreSQL database, use: "postgresql://db_user:db_password@127.0.0.1:5432/db_name?serverVersion=11&charset=utf8"
 # IMPORTANT: You MUST configure your server version, either here or in config/packages/doctrine.yaml
-DATABASE_URL=mysql://db_user:db_password@127.0.0.1:3306/db_name?serverVersion=5.7
+DATABASE_URL=mysql://root:@127.0.0.1:3306/csvdata?serverVersion=5.7
 ###< doctrine/doctrine-bundle ###
