+++ composer.json
@@ -11,6 +11,7 @@
         "symfony/flex": "^1.3.1",
         "symfony/framework-bundle": "5.0.*",
         "symfony/orm-pack": "^1.0",
+        "symfony/twig-bundle": "5.0.*",
         "symfony/yaml": "5.0.*"
     },
     "require-dev": {
