+++ config/packages/test/twig.yaml
@@ -1,0 +1,2 @@
+twig:
+    strict_variables: true
