+++ config/services.yaml
@@ -25,3 +25,5 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
+
+    App\Service\CSVImportWorker:
