+++ symfony.lock
@@ -248,6 +248,26 @@
     "symfony/stopwatch": {
         "version": "v5.0.2"
     },
+    "symfony/translation-contracts": {
+        "version": "v2.0.1"
+    },
+    "symfony/twig-bridge": {
+        "version": "v5.0.2"
+    },
+    "symfony/twig-bundle": {
+        "version": "5.0",
+        "recipe": {
+            "repo": "github.com/symfony/recipes",
+            "branch": "master",
+            "version": "5.0",
+            "ref": "fab9149bbaa4d5eca054ed93f9e1b66cc500895d"
+        },
+        "files": [
+            "config/packages/test/twig.yaml",
+            "config/packages/twig.yaml",
+            "templates/base.html.twig"
+        ]
+    },
     "symfony/var-dumper": {
         "version": "v5.0.2"
     },
@@ -256,6 +276,9 @@
     },
     "symfony/yaml": {
         "version": "v5.0.2"
+    },
+    "twig/twig": {
+        "version": "v3.0.1"
     },
     "zendframework/zend-code": {
         "version": "3.4.1"
